moving openTicket service to services folder. Also updated all components using this service to reflect path change
