+++ src/app/app.module.ts
@@ -24,7 +24,7 @@ import { AdminhistoryComponent } from './adminhistory/adminhistory.component';
 import { SupdateItemComponent } from './supdate-item/supdate-item.component';
 
 import { AuthService } from './services/auth.service';
-import { OpenTicketService } from './shared/opentickets.service';
+import { OpenTicketService } from './services/opentickets.service';
 import { AuthGuard } from './services/auth-guard.service';
 //import { AdminGuard } from './services/admin-guard.service';
 import { CreateTicketService } from './services/create-ticket.service';
